refactor: moved DateUtils into the `internal` package, and converted it to Kotlin
