+++ bugsnag-plugin-react-native/src/main/java/com/bugsnag/android/BreadcrumbDeserializer.java
@@ -1,5 +1,7 @@
 package com.bugsnag.android;
 
+import com.bugsnag.android.internal.DateUtils;
+
 import java.util.Locale;
 import java.util.Map;
