+++ bugsnag-plugin-react-native/src/main/java/com/bugsnag/android/BreadcrumbSerializer.kt
@@ -1,5 +1,6 @@
 package com.bugsnag.android
 
+import com.bugsnag.android.internal.DateUtils
 import java.util.Locale
 
 internal class BreadcrumbSerializer : MapSerializer<Breadcrumb> {
